requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
aiohttp==3.8.6
schedule==1.2.1
httpx==0.25.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
email-validator==2.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
bcrypt==4.0.1 